LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/target-pseudo-in-has-expected.txt

parent color must be yellow green when containing :target
link to #fragment link to #

FAIL CSS Selectors Invalidation: target pseudo in :has() argument assert_equals: parent should be yellowgreen on fragment click expected "rgb(154, 205, 50)" but got "rgb(0, 128, 0)"
